Add stub build command for etest-build.
We don't want to use a shell script to manage such a core piece of the
packaging and want something mildly testable.  This stubs a python click
command to build Docker images for etest that we can re-use in other
places if desired and build a multitude of images (this could easily be
reused in Google cloud builds for example.

Note this is only a stub and the logic still needs to either be ported
from the shell script or added.  Let me know if you'd like me to just
add that functionality in this pull request (seems like a fine idea to
me), but I will need review that I'm converting the logic correctly as
BASH is finicky and quite easy to misinterpret.